Extract MCP server code
