+++ e2e/mcp_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/docker/cagent/cmd/root"
+	"github.com/docker/cagent/pkg/mcp"
 	"github.com/docker/cagent/pkg/teamloader"
 )
 
@@ -22,8 +22,8 @@ func TestMCP_SingleAgent(t *testing.T) {
 		require.NoError(t, team.StopToolSets(ctx))
 	})
 
-	handler := root.CreateToolHandler(team, "root", "testdata/basic.yaml")
-	_, output, err := handler(ctx, nil, root.ToolInput{
+	handler := mcp.CreateToolHandler(team, "root", "testdata/basic.yaml")
+	_, output, err := handler(ctx, nil, mcp.ToolInput{
 		Message: "What is 2+2? Answer in one sentence.",
 	})
 
@@ -43,8 +43,8 @@ func TestMCP_MultiAgent(t *testing.T) {
 		require.NoError(t, team.StopToolSets(ctx))
 	})
 
-	handler := root.CreateToolHandler(team, "web", "testdata/multi.yaml")
-	_, output, err := handler(ctx, nil, root.ToolInput{
+	handler := mcp.CreateToolHandler(team, "web", "testdata/multi.yaml")
+	_, output, err := handler(ctx, nil, mcp.ToolInput{
 		Message: "Say hello in one sentence.",
 	})
